requirements.txt
streamlit
langchain
langchain_openai
selenium
beautifulsoup4
lxml
html5lib
python-dotenv